refactor demo code to extract managers to a separate file
